move model.event -> entity
