+++ entity/benchmark_test.go
@@ -3,6 +3,7 @@ package entity
 import (
 	"github.com/sunist-c/genius-invokation-simulator-backend/entity/model"
 	"github.com/sunist-c/genius-invokation-simulator-backend/enum"
+	"github.com/sunist-c/genius-invokation-simulator-backend/model/context"
 	"testing"
 )
 
@@ -53,5 +54,53 @@ func BenchmarkTestPlayerChainNextWithComplete(b *testing.B) {
 		if !exist {
 			break
 		}
+	}
+}
+
+func BenchmarkTestEventMapPreview(b *testing.B) {
+	m := model.NewEventMap()
+	for i := 0; i < 128; i++ {
+		m.AddEvent(newTestEvent(uint(i), enum.AfterAttack, true, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) }))
+	}
+	m.AddEvent(newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
+	ctx := context.NewCallbackContext()
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		m.Preview(enum.AfterSwitch, ctx)
+	}
+}
+
+func BenchmarkTestEventMapExecute(b *testing.B) {
+	m := model.NewEventMap()
+	for i := 0; i < 128; i++ {
+		m.AddEvent(newTestEvent(uint(i), enum.AfterAttack, true, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) }))
+	}
+	m.AddEvent(newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
+	ctx := context.NewCallbackContext()
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		m.Call(enum.AfterSwitch, ctx)
+	}
+}
+
+func BenchmarkTestEventMapAdd(b *testing.B) {
+	m := model.NewEventMap()
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		m.AddEvent(newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
+	}
+}
+
+func BenchmarkTestEventMapRemove(b *testing.B) {
+	m := model.NewEventMap()
+	event := newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) })
+	m.AddEvent(event)
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		m.RemoveEvent(event)
 	}
 }
