+++ entity/model/event.go
@@ -1,4 +1,4 @@
-package event
+package model
 
 import (
 	"github.com/sunist-c/genius-invokation-simulator-backend/enum"
