+++ entity/unit_test.go
@@ -3,6 +3,7 @@ package entity
 import (
 	"github.com/sunist-c/genius-invokation-simulator-backend/entity/model"
 	"github.com/sunist-c/genius-invokation-simulator-backend/enum"
+	"github.com/sunist-c/genius-invokation-simulator-backend/model/context"
 	"testing"
 )
 
@@ -154,11 +155,15 @@ type testCard struct {
 	t  enum.CardType
 }
 
-func (t testCard) ID() uint { return t.id }
+func (t *testCard) TypeID() uint { return t.id }
 
-func (t testCard) Type() enum.CardType { return t.t }
+func (t *testCard) InjectTypeID(id uint) { t.id = id }
 
-func newTestCard(id uint, t enum.CardType) model.Card { return testCard{id: id, t: t} }
+func (t *testCard) Type() enum.CardType { return t.t }
+
+func (t *testCard) Cost() map[enum.ElementType]uint { return map[enum.ElementType]uint{} }
+
+func newTestCard(id uint, t enum.CardType) model.Card { return &testCard{id: id, t: t} }
 
 func TestCardDeckGet(t *testing.T) {
 	food := newTestCard(1, enum.CardFood)
@@ -411,4 +416,76 @@ func TestPlayerChainNext(t *testing.T) {
 			}
 		}
 	})
+}
+
+type testEvent struct {
+	id      uint
+	trigger enum.TriggerType
+	enable  bool
+	handler func(*context.CallbackContext)
+}
+
+func (t testEvent) ID() uint { return t.id }
+
+func (t testEvent) TriggerAt() enum.TriggerType { return t.trigger }
+
+func (t testEvent) CanTriggered(callbackContext context.CallbackContext) bool { return t.enable }
+
+func (t testEvent) NeedClear() bool { return true }
+
+func (t testEvent) Callback(ctx *context.CallbackContext) { t.handler(ctx) }
+
+func newTestEvent(id uint, trigger enum.TriggerType, enable bool, handler func(callbackContext *context.CallbackContext)) model.Event {
+	return &testEvent{id: id, trigger: trigger, enable: enable, handler: handler}
+}
+
+func TestEventMapExecute(t *testing.T) {
+	switchCharacterEnabled := newTestEvent(0, enum.AfterAttack, true, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) })
+	changeOperatedEnabled := newTestEvent(1, enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) })
+	switchCharacterDisabled := newTestEvent(2, enum.AfterAttack, false, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) })
+	changeOperatedDisabled := newTestEvent(3, enum.AfterSwitch, false, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) })
+
+	testCases := []struct {
+		name      string
+		triggers  []model.Event
+		call      enum.TriggerType
+		judgeFunc func(*context.CallbackContext) bool
+		want      bool
+	}{
+		{
+			name:     "TestEventMapExecute-1",
+			triggers: []model.Event{switchCharacterEnabled, changeOperatedEnabled},
+			call:     enum.AfterAttack,
+			judgeFunc: func(ctx *context.CallbackContext) bool {
+				status, target := ctx.SwitchCharacterResult()
+				return status && target == 114514
+			},
+			want: true,
+		},
+		{
+			name:     "TestEventMapExecute-2",
+			triggers: []model.Event{switchCharacterDisabled, changeOperatedDisabled},
+			call:     enum.AfterAttack,
+			judgeFunc: func(ctx *context.CallbackContext) bool {
+				status, target := ctx.SwitchCharacterResult()
+				return status && target == 114514
+			},
+			want: false,
+		},
+	}
+
+	for _, tt := range testCases {
+		t.Run(tt.name, func(t *testing.T) {
+			m := model.NewEventMap()
+			for _, trigger := range tt.triggers {
+				m.AddEvent(trigger)
+			}
+
+			ctx := context.NewCallbackContext()
+			m.Call(tt.call, ctx)
+			if got := tt.judgeFunc(ctx); got != tt.want {
+				t.Errorf("incorrect execute result: got %v, want %v", got, tt.want)
+			}
+		})
+	}
 }
